refactor: usePortal hook 공통 hooks 하위로 이동
